requirements.txt

tencentcloud-sdk-python==3.0.644
python-dotenv

reportlab